+++ styles/singin.module.scss
@@ -36,34 +36,6 @@
   margin-left: 8px;
 }
 
-// h3 {
-//   color: black;
-//   font-size: 14px;
-//   font-family: Pretendard;
-//   font-weight: 400;
-//   margin-bottom: 12px;
-// }
-
-// label {
-//   color: black;
-//   font-family: Pretendard;
-//   font-size: 14px;
-//   font-weight: 400;
-// }
-
-// input {
-//   width: 400px;
-//   padding: 18px 15px;
-//   border-radius: 8px;
-//   border: 1px #ccd5e3 solid;
-//   color: #373740;
-//   font-size: 16px;
-//   font-family: Pretendard;
-//   font-weight: 400;
-//   line-height: 24px;
-//   margin-top: 12px;
-// }
-
 .hide {
   display: none;
 }
